Extract native gateway component contracts
