+++ exp/runtime/gateway/native_bridge.py
@@ -28,19 +28,16 @@ from __future__ import annotations
 import json
 import threading
 import time
-from collections.abc import Callable, Mapping
+from collections.abc import Callable
 from dataclasses import dataclass, field, replace
-from typing import Protocol, cast
+from typing import cast
 
 from exp.common.core.artifacts import JsonObject, stable_id
-from exp.common.models.gateway_catalog import ExactModelDeployment
 from exp.runtime.gateway.boundary import boundary_protocol_error
 from exp.runtime.gateway.budgets import BudgetReservationRejected, maximum_attempt_cost_micro_usd
 from exp.runtime.gateway.contracts import (
-    AttemptId,
     AuthorizationSnapshot,
     DirectTarget,
-    ExecutionSnapshot,
     GatewayApiSurface,
     GatewayEvent,
     GatewayEventKind,
@@ -60,21 +57,19 @@ from exp.runtime.gateway.discovery import (
 # the native path must enforce the same one, so the private helper is shared.
 from exp.runtime.gateway.execution import (
     GatewayExecutionError,
-    GatewayExecutor,
     _require_deployment_identity,  # noqa: PLC2701
 )
 from exp.runtime.gateway.group_commit import SyncGroupCommitLedger
-from exp.runtime.gateway.interfaces import GatewayControlStore
 from exp.runtime.gateway.ledger import SQLiteAttemptLedger
+from exp.runtime.gateway.native_components import NativeGatewayComponents
 from exp.runtime.gateway.native_responses import (
     ContinuationContext,
     continued_request,
     remember_turn,
     responses_envelope,
 )
-from exp.runtime.gateway.routing import CatalogRouteResolver, GatewayRoute, GatewayRoutingError
+from exp.runtime.gateway.routing import GatewayRoute, GatewayRoutingError
 from exp.runtime.gateway.usage import GatewayUsageReport, read_usage_report, usage_html
-from exp.runtime.models import RuntimeModelCatalog
 from exp.runtime.models.providers import (
     preflight_gateway_request,
     require_gateway_provider,
@@ -105,92 +100,6 @@ _TERMINAL_KINDS = {
     "incomplete": GatewayEventKind.INCOMPLETE,
     "failed": GatewayEventKind.FAILED,
 }
-
-
-class NativeAttemptLedger(Protocol):
-    """Synchronous durable ledger used from native callback threads."""
-
-    def accept_request(self, *, authorization: AuthorizationSnapshot) -> None:
-        """Persist one accepted request."""
-        ...
-
-    def start_attempt(
-        self,
-        *,
-        snapshot: ExecutionSnapshot,
-        deployment: ExactModelDeployment,
-        attempt_ordinal: int,
-        route_depth: int,
-        maximum_cost_micro_usd: int | None = None,
-        route_reason: str | None = None,
-        fallback_reason: str | None = None,
-    ) -> AttemptId:
-        """Reserve and persist one provider attempt."""
-        ...
-
-    def finish_attempt(
-        self,
-        *,
-        attempt_id: AttemptId,
-        terminal_event: GatewayEvent | None,
-        failure: GatewayFailure | None,
-        finalize_request: bool = True,
-    ) -> None:
-        """Settle one provider attempt."""
-        ...
-
-    def finish_request(
-        self,
-        *,
-        authorization: AuthorizationSnapshot,
-        failure: GatewayFailure,
-    ) -> None:
-        """Finalize accepted work that failed before dispatch."""
-        ...
-
-
-class NativeGatewayComponents(Protocol):
-    """Engine-neutral components required by the native control plane."""
-
-    @property
-    def store(self) -> GatewayControlStore:
-        """Return the authority store."""
-        ...
-
-    @property
-    def ledger(self) -> NativeAttemptLedger:
-        """Return the synchronous durable accounting ledger."""
-        ...
-
-    @property
-    def routes(self) -> CatalogRouteResolver:
-        """Return the direct-route resolver."""
-        ...
-
-    @property
-    def executor(self) -> GatewayExecutor:
-        """Return the shared accounting-health latch."""
-        ...
-
-    @property
-    def reconciled_expired_requests(self) -> int:
-        """Return startup-reconciled request count."""
-        ...
-
-    @property
-    def reconciled_unknown_attempts(self) -> int:
-        """Return startup-reconciled attempt count."""
-        ...
-
-    @property
-    def runtime_catalogs(self) -> Mapping[tuple[str, str], RuntimeModelCatalog]:
-        """Return runtime catalogs keyed by alias revision and digest."""
-        ...
-
-    @property
-    def organization_id(self) -> str:
-        """Return the organization used by the local usage endpoint."""
-        ...
 
 
 class _NativeDialectUnavailableError(RuntimeError):
